+++ src/ace_time/hw/StmRtc.h
@@ -1,10 +1,10 @@
 /*
  * MIT License
- * Copyright (c) 2018 Brian T. Park, Anatoli Arkhipenko
+ * Copyright (c) 2020 Brian T. Park, Anatoli Arkhipenko
  */
 
-#ifndef ACE_TIME_HW_STMRTC_H
-#define ACE_TIME_HW_STMRTC_H
+#ifndef ACE_TIME_HW_STM_RTC_H
+#define ACE_TIME_HW_STM_RTC_H
 
 #if ! defined(UNIX_HOST_DUINO)
 #if defined(ARDUINO_ARCH_STM32)
@@ -18,11 +18,11 @@ namespace hw {
 class HardwareDateTime;
 
 /**
- * A class that reads and writes HardwareDateTime and HardwareTemperature from a
- * STM RTC chip. This class is designed to access just enough features of the
- * RTC chip to implement the ace_time::StmRtcClock class. It is not
- * meant to provide access to all the features of the RTC chip.
- 
+ * A class that reads and writes HardwareDateTime from a STM RTC chip. This
+ * class is designed to access just enough features of the RTC chip to
+ * implement the ace_time::StmRtcClock class. It is not meant to provide access
+ * to all the features of the RTC chip.
+ *
  * Requires https://github.com/stm32duino/STM32RTC
  */
 class StmRtc {
@@ -36,15 +36,16 @@ class StmRtc {
     /** Set the STM with the HardwareDateTime values. */
     void setDateTime(const HardwareDateTime& dateTime) const;
 
+    /** Return true if the RTC is available and the time is set. */
     bool isTimeSet() const;
+
   private:
     STM32RTC* mRtc;
 };
 
-}
-}
-#endif  //  #if defined(ARDUINO_ARCH_STM32)
+} // hw
+} // ace_time
 
-#endif  //  #if ! defined(UNIX_HOST_DUINO)
-
-#endif  //  #ifndef ACE_TIME_HW_STMRTC_H
+#endif //  #if defined(ARDUINO_ARCH_STM32)
+#endif //  #if ! defined(UNIX_HOST_DUINO)
+#endif //  #ifndef ACE_TIME_HW_STM_RTC_H